Refactored main logic into seperate libary project to more easily unit test
