+++ ADF.Functions/ADF.Functions.csproj
@@ -18,11 +18,11 @@
     <RootNamespace>Azure.Datafactory.Extensions.Functions</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.2.2" />
-    <PackageReference Include="Azure.Storage.Files.DataLake" Version="12.4.0" />
-    <PackageReference Include="Flurl" Version="2.8.2" />
-    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="3.0.3" />
-    <PackageReference Include="System.Linq.Dynamic.Core" Version="1.2.2" />
+    <PackageReference Include="Azure.Identity" Version="1.3.0" />
+    <PackageReference Include="Azure.Storage.Files.DataLake" Version="12.6.0" />
+    <PackageReference Include="Flurl" Version="3.0.1" />
+    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="3.0.11" />
+    <PackageReference Include="System.Linq.Dynamic.Core" Version="1.2.8" />
   </ItemGroup>
   <ItemGroup>
     <None Update="host.json">
@@ -36,5 +36,8 @@
       <Pack>True</Pack>
       <PackagePath></PackagePath>
     </None>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Azure.DataFactory.Extensions\Azure.DataFactory.Extensions.csproj" />
   </ItemGroup>
 </Project>
