+++ Azure.DataFactory.Extensions/Azure.DataFactory.Extensions.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+    <OutDir>bin\$(Configuration)</OutDir>
+    <Version>1.0.0</Version>
+    <Company />
+    <Authors>Niall Langley</Authors>
+    <Product>azure.datafactory.extensions</Product>
+    <PackageId>Azure.Datafactory.Extensions</PackageId>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <PackageProjectUrl>https://github.com/NJLangley/azure.datafactory.extensions</PackageProjectUrl>
+    <AssemblyName>Azure.Datafactory.Extensions</AssemblyName>
+    <RootNamespace>Azure.Datafactory.Extensions</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Azure.Identity" Version="1.3.0" />
+    <PackageReference Include="Azure.Storage.Files.DataLake" Version="12.6.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.13" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="System.Linq.Dynamic.Core" Version="1.2.8" />
+  </ItemGroup>
+
+</Project>
